Rename ov16_02266F1C -> healthbar
