+++ platinum.us/main.lsf
@@ -699,7 +699,7 @@ Overlay overlay16
 	Object main.nef.p/src_overlay016_battle_display.c.o
 	Object main.nef.p/src_overlay016_ov16_02264798.c.o
 	Object main.nef.p/src_overlay016_battle_io.c.o
-	Object main.nef.p/src_overlay016_ov16_02266F1C.c.o
+	Object main.nef.p/src_overlay016_healthbar.c.o
 	Object main.nef.p/src_overlay016_ov16_02268520.c.o
 	Object main.nef.p/src_overlay016_ov16_0226871C.c.o
 	Object main.nef.p/src_overlay016_party_gauge.c.o
