+++ src/overlay016/ov16_0223DF00.c
@@ -76,7 +76,7 @@
 #include "overlay016/ov16_0223DF00.h"
 #include "battle/battle_lib.h"
 #include "battle/battle_display.h"
-#include "overlay016/ov16_02266F1C.h"
+#include "battle/healthbar.h"
 #include "overlay016/ov16_02268520.h"
 #include "overlay016/ov16_0226871C.h"
 #include "overlay016/ov16_0226E148.h"
